% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlite_id.R
\name{create_sqlite_id}
\alias{create_sqlite_id}
\title{Function to create id_cols table within database}
\usage{
create_sqlite_id(conn = conn, data = gov_census)
}
\arguments{
\item{conn}{Database connection}

\item{data}{data.table from Willamette (gov_fin_data.csv)}
}
\description{
Selects unchanging identifier columns including name, fips identifiers
There are no params given, but gov_census must be in global env
The name of table to "id_cols" and index is set to "id" automatically
}
\examples{
\dontrun{
create_identifiers()}

}
