% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chart_data.R
\name{get_chart_data}
\alias{get_chart_data}
\title{Function takes variable arguments and returns data for table}
\usage{
get_chart_data(conn, table, fields)
}
\arguments{
\item{conn}{Database connection}

\item{table}{Name of table}

\item{fields}{Vector of all fields}
}
\description{
With input of table and fields, returns a data.frame
}
\examples{
\dontrun{
get_chart_data(
  conn = conn,
  table = table,
  fields = c("total_revenue", "total_taxes"))}

}
