% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_clean_govcensus_csv.R
\name{load_and_clean_govcensus_csv}
\alias{load_and_clean_govcensus_csv}
\title{Function to load downloaded, unpacked .csv and prepare for database}
\usage{
load_and_clean_govcensus_csv(path = "inst/extdata")
}
\arguments{
\item{path}{Location to store data}
}
\description{
Loads downloaded .csv as data.table, cleans names,
}
\examples{
\dontrun{
load_and_clean_govcensus_csv(path = "data")}

}
