% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_db.R
\name{make_db}
\alias{make_db}
\title{Function builds SQLite database from Census Type Codes (ie: Federal, State, etc)}
\usage{
make_db(data = gov_census_list, connect = "duckdb")
}
\arguments{
\item{data}{Census data.table}

\item{connect}{Type of connection (duckdb or sqlite)}
}
\description{
Loads the .csv download as data.table, separates identifying columns
uses make_sqlite_table() to add tables for each type code including:
fedl, state, local, special and school.
Finally adds table for identifier columns with create_sqlite_id(), closes conn and deletes .csv
}
\examples{
\dontrun{
make_sqlite_db())}

}
