% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sqlite_table.R
\name{make_sqlite_table}
\alias{make_sqlite_table}
\title{Function to subset of gov_census and loads into database}
\usage{
make_sqlite_table(conn = conn, data = gov_census, type, name, index = "id")
}
\arguments{
\item{conn}{Database connection}

\item{data}{data.table from Willamette (gov_fin_data.csv)}

\item{type}{"Type Code" chr from Williamette data subsets ("0", "1", "2", "3, "4, "5", "6")}

\item{name}{Name chr to give to table ("fedl", "local", etc)}

\item{index}{Usually "id" variable common to all data subsets}
}
\description{
Given table type_code, name and database index (if required),
Selects non zero columns, filters and creates table
Must have data in global env and database connection
}
\examples{
\dontrun{
make_sqlite_table(type = c("2", "3"), name = "local", index = "id")}

}
