% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Function to subset of gov_census and loads into database}
\usage{
make_table(conn = conn, name, data = table)
}
\arguments{
\item{conn}{Database connection}

\item{name}{Name chr to give to table ("fedl", "local", etc)}

\item{data}{data.table from Willamette (gov_fin_data.csv)}
}
\description{
Given table type_code, name and database index (if required),
Selects non zero columns, filters and creates table
Must have data in global env and database connection
}
\examples{
\dontrun{
make_sqlite_table(conn = conn, name = "local", data = table)}

}
