% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_govfin_data.R
\name{unzip_govfin_data}
\alias{unzip_govfin_data}
\title{Function to extract and rename .csv from downloaded zipped folder}
\usage{
unzip_govfin_data(path = "inst/extdata")
}
\arguments{
\item{path}{Folder in working dir where zip is stored}
}
\description{
Unzip folder as .csv but does not extract other documentation files
}
\examples{
\dontrun{
unzip_govfin_data(path = "data") }

}
