% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_gov_census_db_wrapper.R
\name{full_gov_census_db_wrapper}
\alias{full_gov_census_db_wrapper}
\title{Wrapper for all full data prep process}
\usage{
full_gov_census_db_wrapper()
}
\description{
Runs get_govfin_data(), unzip_govfin_data(), load_and_clean_govcensus_csv(),
then make_sqlite_db() builds database in inst/extdata with
make_sqlite_table() and create_sqlite_id()
}
\examples{
\dontrun{
load_and_clean_govcensus_csv(path = "data")}

}
